Conflicts:
	src/lib/analysis/Util.cpp
	src/lib/banal/StructSimple.cpp
	src/lib/prof-lean/hpcrun-fmt.h
	src/lib/prof/CCT-Tree.hpp
	src/lib/prof/CallPath-Profile.cpp
	src/lib/prof/Metric-ADesc.hpp
	src/lib/prof/Metric-Mgr.cpp
	src/lib/prof/NameMappings.cpp
	src/lib/xml/hpc-experiment.dtd.h
	src/lib/xml/hpc-structure.dtd.h
	src/tool/hpcfnbounds/Makefile.in
	src/tool/hpcfnbounds/main.cpp
	src/tool/hpcrun/Makefile.am
	src/tool/hpcrun/Makefile.in
	src/tool/hpcrun/cct/cct.c
	src/tool/hpcrun/cct/cct.h
	src/tool/hpcrun/cct/cct_bundle.c
	src/tool/hpcrun/cct/cct_bundle.h
	src/tool/hpcrun/cct_insert_backtrace.c
	src/tool/hpcrun/fnbounds/fnbounds_client.c
	src/tool/hpcrun/fnbounds/fnbounds_dynamic.c
	src/tool/hpcrun/metrics.c
	src/tool/hpcrun/metrics.h
	src/tool/hpcrun/ompt/ompt-callstack.c
	src/tool/hpcrun/sample-sources/perf/event_custom.h
	src/tool/hpcrun/sample-sources/perf/kernel_blocking.c
	src/tool/hpcrun/sample-sources/perf/linux_perf.c
	src/tool/hpcrun/sample-sources/perf/perf-util.c
	src/tool/hpcrun/sample-sources/perf/perf-util.h
	src/tool/hpcrun/sample_event.c

+++ src/lib/support/VarMap.hpp
@@ -12,7 +12,7 @@
 // HPCToolkit is at 'hpctoolkit.org' and in 'README.Acknowledgments'.
 // --------------------------------------------------------------------------
 //
-// Copyright ((c)) 2002-2019, Rice University
+// Copyright ((c)) 2002-2020, Rice University
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -57,12 +57,12 @@ private:
   hpcrun_metricVal_t *m_metrics;	   // metric values
   metric_desc_t*      m_list_metric_desc;  // metric descriptions
 
-  size_t m_num_metrics;
+  int m_num_metrics;
   int    m_error_code;
 
 public:
   bool   isVariable(char *expr);
-  double getValue(unsigned int var);
+  double getValue(int var);
   int    getErrorCode();
 
   VarMap(hpcrun_metricVal_t *metrics,
